+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [0.5.92] - 2020-02-11
+### Added
+- The `add_separators()` function.
+- The `session error redirect url` Configuration directive.
+### Changed
+- Calling `as_noun()` on a group with one item will return the
+  `noun_singular()` of the applicable noun.
+
 ## [0.5.91] - 2020-02-07
 ### Added
 - The `/api/interview_data` API endpoint.